Refactor tweets.go to improve error handling and logging
The `tweets.go` file was refactored to streamline the error handling process and incorporate improved logging for the `getTweetTx` function. This includes the addition of an error log in the event of an issue when creating a transaction. Transaction commitments have also been relocated to the main function for optimal error management. Furthermore, a predefined `bufferSize` was implemented to replace the numerous hardcoded buffer size values throughout the code.