Add error check to NewTransaction in tweets.go
An additional error check has been added to the NewTransaction function in tweets.go. This check returns an error immediately if there is an issue with the creation of a new transaction. This ensures that the program doesn't call other functions or uses resources when it's unnecessary to do so.