Update migration version check and transaction execution in db migration
Refined the version check in the migration process and made alterations in db migration file with updates in transaction execution cycle. The Version() function in 00-init.go is tweaked to return a uint32 value in line with the changes. Also, a new migration test file has been introduced to validate the updates.