Implement single linked list for search request task management
A single linked list data structure has been added to the project to concurrently manage search request tasks. This improves how tasks are pushed and popped in the watcher. The code related to searching tweets has also been refactored to enhance performance and readability. The modification includes error handling updates and changes in the way search cursors are initialized and operated on search requests.