Implement tweet counting mechanism in Twitter scrapper
Added a mechanism to count processed tweets in the Twitter scrapper tool. This allows us to track the number of tweets processed over time for performance and monitoring purposes. To achieve this, an atomic counter was introduced in the repository layer and data gathering function was added in the metrics layer.