Implement tweetseditor package for crypto-tweet-sense tool
This update introduces a new tweetseditor package to facilitate the management of tweets within the crypto-tweet-sense tool. It provides an interface for editing and update operations, as well as channels to subscribe for edited tweets. The manager struct in the code base is now equipped with sendInterval, cleanInterval attributes, and methods for editing and subscribing for Tweet changes. Additionally, test fixtures for tweets were added.