Migrate from FoundationDB to Redis and deprecate old interfaces
This commit majorly consists of transitioning from FoundationDB to Redis in various parts of the code. Also, as part of this migration, some of the interfaces related to FoundationDB have been deprecated. Furthermore, changes have been made in order to integrate fetching with session repositories. To ensure a smooth transition, migration scripts have been included.