Add database migration function to twitter scrapper
Added a new "Migrator" function to the twitter scrapper application. It aims to streamline the process of migrating database versions. The Migrator function monitors the current version and performs necessary updates to meet the required version. This will make the application more robust to changes in the structure or type of the stored data.