Added Empty migration type to migration.go
A new struct, Empty, was added to the migration.go file. It contains three functions: Up, Down, and Version. The Up and Down functions both return nil, representing a no-operation (no-op) migration type, and the Version function returns 1. An instance of this new struct was also added to the Migrations function.