Remove escaping content in story summary output
Removed the utility function for escaping the content of a generated story summary in two methods of 'editor.go'. Also, added a new migration for SessionStorage in the 'migration.go' and '03-session-storage.go'. Additionally, a new host 'grafana.lueurxax.org' has been included in the 'values.yaml' for Grafana ingress.